zabbix module renamed to pyzabbix
