+++ pyzabbix/__init__.py
@@ -1,0 +1,6 @@
+import pyzabbix.version
+
+from .api import ZabbixAPI, ZabbixAPIException
+from .sender import ZabbixMetric, ZabbixSender
+
+__version__ = pyzabbix.version.__version__
