--- zabbix/__init__.py
@@ -1,3 +1,0 @@
-import zabbix.version
-
-__version__ = zabbix.version.__version__
